Revert "fix EventIndex() overflow by using int64_t"

+++ event.h
@@ -36,6 +36,6 @@ void EventFree(EVENT *ev);
 int EventSetCause(EVENT *ev, unsigned long cause_host,
 				  unsigned long long cause_seq_no);
 
-int64_t EventIndex(unsigned long host, unsigned long long seq_no);
+double EventIndex(unsigned long host, unsigned long long seq_no);
 
 #endif
